Make all Parcelable classes public
Since Robolectric may run on JDK 8, it means that there is an extra
security check when accessing class field using reflection:
On Android, the VM will only check if the field is set as public when accessing
it via reflect, but on JDK 8, it will also make sure that the caller class (in our
case android.os.Parcel) can access the class declaring the field.
So, if we have a Parcelable that is non-public, under JDK 8 Parcel may not have
access to read CREATOR static field!

This change makes all Parcelable classes public, hence circumvent this difference between
Android's VM and JDK 8.